server build ok

+++ src/base/file_util.cc
@@ -33,8 +33,10 @@ bool Stat(const std::string& path, struct stat* st) {
 }
 
 void RemoveFile(const std::string& path) {
-  int ret = ::unlink(path.c_str());
-  PLOG_IF(WARNING, ret != 0) << "unlink failed: " << path;
+  if (FileExist(path)) {
+    int ret = ::unlink(path.c_str());
+    PLOG_IF(WARNING, ret != 0) << "unlink failed: " << path;
+  }
 }
 
 bool FileExist(const std::string& path) {
